requirements.txt
streamlit
google-generativeai
python-docx
langchain-google-genai
langchain
langchain-community
langchain-text-splitters
faiss-cpu